remove finished animations
before this fix finished floating animations would still be floating around and stealing resources but outside of the screen, if you do this with enough floaters, you run in some problems (50+ is enough to start fans etc). The child gets removed from the DOM after the `float` animation is finished.

fixes #2